DOES NOT WORK -- Tried to implement a CategoryListTile; not sure why ListTile is causing the error
